修正1回目:他者が削除できないように設定

+++ app/controllers/items_controller.rb
@@ -39,8 +39,10 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    if @item.destroy
-      redirect_to root_path
+    if current_user.id == @item.user_id
+      if @item.destroy
+        redirect_to root_path
+      end
     end
   end
